move walker to another file
