+++ Composer/packages/lib/shared/src/copyUtils.ts
@@ -1,43 +1,6 @@
 import cloneDeep from 'lodash.clonedeep';
 
-const NestedFieldNames = {
-  Actions: 'actions',
-  ElseActions: 'elseActions',
-  DefaultCase: 'default',
-  Cases: 'cases',
-};
-
-const DEFAULT_CHILDREN_KEYS = [NestedFieldNames.Actions];
-const childrenMap = {
-  ['Microsoft.IfCondition']: [NestedFieldNames.Actions, NestedFieldNames.ElseActions],
-  ['Microsoft.SwitchCondition']: [NestedFieldNames.Cases, NestedFieldNames.DefaultCase],
-};
-
-/**
- * Considering that an Adaptive Action could be nested with other actions,
- * for example, the IfCondition and SwitchCondition and Foreach, we need
- * this helper to visit all possible action nodes recursively.
- *
- * @param {any} input The input Adaptive Action which has $type field.
- * @param {function} visitor The callback function called on each action node.
- */
-async function walkAdaptiveAction(input: any, visitor: (data: any) => Promise<any>) {
-  if (!input || !input.$type) return;
-
-  await visitor(input);
-
-  let childrenKeys = DEFAULT_CHILDREN_KEYS;
-  if (input.$type && childrenMap[input.$type]) {
-    childrenKeys = childrenMap[input.$type];
-  }
-
-  for (const childrenKey of childrenKeys) {
-    const children = input[childrenKey];
-    if (Array.isArray(children)) {
-      Promise.all(children.map(async x => await walkAdaptiveAction(x, visitor)));
-    }
-  }
-}
+import { walkAdaptiveAction } from './walkAdaptiveAction';
 
 const overrideLgActivity = async (data, { copyLgTemplate }) => {
   const newLgId = `[bfdactivity-${data.$designer.id}]`;
